+++ tests/commander/test_handler.py
@@ -27,8 +27,8 @@ def test_add_callback_functions(handler_coroutine: HandlerCoroutine):
     handler_coroutine.add_callback_functions(functions_info=[function_info, function_info], which="at_handler_start")
 
     # Assert
-    assert handler_coroutine._callback is not None
-    assert handler_coroutine._callback.at_handler_start == [function_info, function_info, function_info]
+    assert handler_coroutine.callback is not None
+    assert handler_coroutine.callback.at_handler_start == [function_info, function_info, function_info]
     with pytest.raises(ValueError):
         handler_coroutine.add_callback_functions(functions_info=function_info, which="error_callback_type")
 
@@ -43,8 +43,8 @@ def test_add_callback(handler_coroutine: HandlerCoroutine):
     handler_coroutine.add_callback([callback, callback])
 
     # Assert
-    assert handler_coroutine._callback is not None
-    assert handler_coroutine._callback.at_handler_start == [function_info, function_info, function_info]
+    assert handler_coroutine.callback is not None
+    assert handler_coroutine.callback.at_handler_start == [function_info, function_info, function_info]
 
 
 def test_call_handler(handler_coroutine: HandlerCoroutine):
@@ -98,3 +98,49 @@ def test_exception_callback(handler_coroutine: HandlerCoroutine, commander: Comm
     assert callback_function.called
     assert manipulate[0] == 1
     assert test_handler.state == "EXCEPTION"
+
+
+class HandlerClass:
+    def __init__(self, obj_list: list):
+        self.obj_list = obj_list
+
+    @handler(PASS_WORD)
+    async def method_handler(self, self_handler, obj_list: list, *args, **kwargs):
+        obj_list[0] += 1
+
+def test_handler_decorator(commander: CommanderAsync):
+    # Setup
+    manipulate = [0]
+    @handler(PASS_WORD)
+    async def nested_handler(self_handler, obj_list: list, *args, **kwargs):
+        obj_list[0] += 1
+    handlerclass = HandlerClass(manipulate)
+    callback_function = Mock()
+    function_info: CallbackDict = {"function": callback_function, "inject_task_node": True}
+    callback_1 = Callback(at_handler_start=[function_info])
+    callback_2 = Callback(at_handler_start=[function_info])
+    callback_3 = Callback(at_handler_start=[function_info])
+    callback_4 = Callback(at_handler_start=[function_info])
+    handler_in_class = handlerclass.method_handler(manipulate, callback_1, callback=callback_2)
+    a_nested_handler = nested_handler(manipulate, callback_3, callback=callback_4)
+
+    # Action
+    threading.Thread(target=commander.run).start()
+    while not commander.running_status:
+        pass
+    commander.call_handler_threadsafe(handler_in_class)
+    commander.call_handler_threadsafe(a_nested_handler)
+    while not commander.is_empty():
+        pass
+    commander.exit()
+
+    # Assert
+    assert manipulate[0] == 2
+    assert callback_function.called
+    call_args_list = callback_function.call_args_list
+    kwargs_values = [
+        value for _, kwargs in call_args_list for value in kwargs.values()
+    ]
+    assert kwargs_values.count(handler_in_class) == 2
+    assert kwargs_values.count(a_nested_handler) == 2
+    assert callback_function.call_count == 4